+++ src/GraphQLIDE.ts
@@ -1,0 +1,5 @@
+export enum GraphQLIDE {
+  None = 'None',
+  GraphiQL = 'GraphiQL',
+  ApolloSandbox = 'ApolloSandbox',
+}