+++ src/GraphqlMockingService.ts
@@ -1,16 +1,21 @@
 import GraphqlMockingContext from './GraphqlMockingContext';
+import {GraphQLIDE} from './GraphQLIDE';
 import MockServer from './MockServer';
 
 type GraphqlMockingServiceOptions = {
   port?: number;
   subgraph?: boolean;
+  graphQLIDE?: GraphQLIDE;
 };
 
 export default class GraphqlMockingService {
   readonly server;
   private subgraph;
   constructor(private options: GraphqlMockingServiceOptions = {}) {
-    this.server = new MockServer({port: options.port});
+    this.server = new MockServer({
+      port: options.port,
+      graphQLIDE: options.graphQLIDE,
+    });
     this.subgraph = options.subgraph || false;
   }