+++ src/MockServer.ts
@@ -1,9 +1,11 @@
 import fetch, {Response} from 'node-fetch';
 import createApp from './utilities/createApp';
 import {Seed} from './seed/types';
+import {GraphQLIDE} from './GraphQLIDE';
 
 type MockServerOptions = {
   port?: number;
+  graphQLIDE?: GraphQLIDE;
 };
 
 type SchemaRegistrationOptions = {
@@ -11,14 +13,16 @@ type SchemaRegistrationOptions = {
 };
 
 class MockServer {
-  readonly port;
+  readonly port: number;
+  readonly graphQLIDE: GraphQLIDE;
   private appServer;
   constructor(options: MockServerOptions) {
     this.port = options.port || 5000;
+    this.graphQLIDE = options.graphQLIDE || GraphQLIDE.ApolloSandbox;
   }
 
   async start(): Promise<void> {
-    const app = createApp();
+    const app = createApp({graphQLIDE: this.graphQLIDE, port: this.port});
 
     this.appServer = await app.listen({port: this.port}, () =>
       console.log(