+++ src/utilities/createApp.ts
@@ -5,11 +5,21 @@ import seedRoutes from '../routes/seed';
 import schemaRoutes from '../routes/schema';
 import SeedManager from '../seed/SeedManager';
 import ApolloServerManager from '../ApolloServerManager';
+import {GraphQLIDE} from '../GraphQLIDE';
 
 /**
+ * @param {object} root0 - The root object
+ * @param {GraphQLIDE} root0.graphQLIDE - The type of GraphQL IDE to use
+ * @param {number} root0.port - The port to run the server on
  * @returns {express.Express} An express server instance
  */
-export default function createApp(): express.Express {
+export default function createApp({
+  graphQLIDE,
+  port,
+}: {
+  graphQLIDE: GraphQLIDE;
+  port: number;
+}): express.Express {
   const app = express();
   const seedManager = new SeedManager();
   const apolloServerManager = new ApolloServerManager();
@@ -30,7 +40,10 @@ export default function createApp(): express.Express {
     return res;
   });
 
-  app.use('/graphql', graphqlRoutes({seedManager, apolloServerManager}));
+  app.use(
+    '/graphql',
+    graphqlRoutes({graphQLIDE, seedManager, apolloServerManager, port})
+  );
   app.use('/schema', schemaRoutes({apolloServerManager}));
   app.use('/seed', seedRoutes({seedManager}));