+++ CHANGELOG.md
@@ -8,6 +8,19 @@ and this project adheres to
 
 ## [Unreleased]
 
+## [v1.1.0] - 2023-06-28
+
+### Added
+
+- feat: add support for serving a GraphQL IDE from the mock servers `/graphql`
+  route. Three options are available:
+  - **DEFAULT** `GraphQLIDE.ApolloSandbox`: Serve's the
+    [Apollo Sandbox](https://www.apollographql.com/docs/graphos/explorer/sandbox/)
+    experience.
+  - `GraphQLIDE.GraphiQL`: Serve's the latest version of
+    [GraphiQL](https://github.com/graphql/graphiql/tree/main/packages/graphiql#readme)
+  - `GraphQLIDE.None`: Disables the GraphQL IDE experience
+
 ### Fixed
 
 - fix: don't require sequenceId when executing a query